fix: validation

+++ backend/src/modules/task/service.js
@@ -1,7 +1,17 @@
 const taskRepository = require("../../database/repositories/taskRepository");
+const { NotFound } = require("../../errors");
 
 exports.getById = async (taskId) => {
   const task = await taskRepository.findById(taskId);
+  if (!task)
+    throw new NotFound("The task does not found", [
+      {
+        type: "field",
+        msg: "the id does not match any record",
+        location: "params",
+        property: "id",
+      },
+    ]);
   return task;
 };
 
@@ -16,11 +26,34 @@ exports.store = async (taskData) => {
 };
 
 exports.update = async (taskId, taskData) => {
+  const isExist = await taskRepository.exists(taskId);
+  if (!isExist)
+    throw new NotFound("The task does not found", [
+      {
+        type: "field",
+        msg: "the id does not match any record",
+        location: "params",
+        property: "id",
+      },
+    ]);
+
   await taskRepository.update(taskId, taskData);
+
   const task = await taskRepository.findById(taskId);
   return task;
 };
 
 exports.delete = async (taskId) => {
+  const isExist = await taskRepository.exists(taskId);
+  if (!isExist)
+    throw new NotFound("The task does not found", [
+      {
+        type: "field",
+        msg: "the id does not match any record",
+        location: "params",
+        property: "id",
+      },
+    ]);
+
   await taskRepository.delete(taskId);
 };
